# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enitan Afolabi & Company API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form submissions
  - name: careers
    description: Career applications
  - name: blog
    description: Blog posts
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactFormInput"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contact/messages:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List all contact messages (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of contact messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessageList"

  /contact/messages/{id}:
    get:
      operationId: getContactMessage
      tags: [contact]
      summary: Get a single contact message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "404":
          description: Not found
    patch:
      operationId: updateContactMessage
      tags: [contact]
      summary: Update contact message status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactMessageInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /careers:
    post:
      operationId: submitCareerApplication
      tags: [careers]
      summary: Submit a career application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CareerApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerApplication"
        "400":
          description: Validation error

  /careers/applications:
    get:
      operationId: listCareerApplications
      tags: [careers]
      summary: List all career applications (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerApplicationList"

  /careers/applications/{id}:
    get:
      operationId: getCareerApplication
      tags: [careers]
      summary: Get a single career application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Career application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerApplication"
        "404":
          description: Not found
    patch:
      operationId: updateCareerApplication
      tags: [careers]
      summary: Update career application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCareerApplicationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerApplication"

  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 9
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPostList"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a new blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogPostInput"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /blog/categories:
    get:
      operationId: listBlogCategories
      tags: [blog]
      summary: List blog categories
      responses:
        "200":
          description: Blog categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogCategoryList"

  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogPostInput"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin user
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required:
        - error

    ContactFormInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
      required:
        - fullName
        - email
        - subject
        - message

    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [new, read, replied]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - subject
        - message
        - status
        - createdAt

    UpdateContactMessageInput:
      type: object
      properties:
        status:
          type: string
          enum: [new, read, replied]

    ContactMessageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContactMessage"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - totalPages

    CareerApplicationInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        coverNote:
          type: string
        cvUrl:
          type: string
      required:
        - fullName
        - email
        - coverNote

    CareerApplication:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        position:
          type: string
        coverNote:
          type: string
        cvUrl:
          type: string
        status:
          type: string
          enum: [new, reviewing, shortlisted, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - coverNote
        - status
        - createdAt

    UpdateCareerApplicationInput:
      type: object
      properties:
        status:
          type: string
          enum: [new, reviewing, shortlisted, rejected]

    CareerApplicationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CareerApplication"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - totalPages

    BlogPost:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        featuredImage:
          type: string
        published:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - excerpt
        - content
        - category
        - author
        - published
        - createdAt

    CreateBlogPostInput:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        featuredImage:
          type: string
        published:
          type: boolean
      required:
        - slug
        - title
        - excerpt
        - content
        - category
        - author

    UpdateBlogPostInput:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        author:
          type: string
        featuredImage:
          type: string
        published:
          type: boolean

    BlogPostList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BlogPost"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - totalPages

    BlogCategoryList:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
      required:
        - categories

    AdminStats:
      type: object
      properties:
        totalMessages:
          type: integer
        newMessages:
          type: integer
        totalApplications:
          type: integer
        newApplications:
          type: integer
        totalBlogPosts:
          type: integer
        publishedBlogPosts:
          type: integer
      required:
        - totalMessages
        - newMessages
        - totalApplications
        - newApplications
        - totalBlogPosts
        - publishedBlogPosts

    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - success

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
      required:
        - id
        - email
        - name