# lib/api-zod/src/generated/types/blogPostList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enitan Afolabi & Company API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlogPost } from "./blogPost";

export interface BlogPostList {
  items: BlogPost[];
  total: number;
  page: number;
  totalPages: number;
}