# lib/api-zod/src/generated/types/careerApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enitan Afolabi & Company API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareerApplicationStatus } from "./careerApplicationStatus";

export interface CareerApplication {
  id: number;
  fullName: string;
  email: string;
  phone?: string;
  position?: string;
  coverNote: string;
  cvUrl?: string;
  status: CareerApplicationStatus;
  createdAt: Date;
}