# lib/api-zod/src/generated/types/careerApplicationList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enitan Afolabi & Company API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareerApplication } from "./careerApplication";

export interface CareerApplicationList {
  items: CareerApplication[];
  total: number;
  page: number;
  totalPages: number;
}