# lib/api-zod/src/generated/types/contactMessageList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enitan Afolabi & Company API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactMessage } from "./contactMessage";

export interface ContactMessageList {
  items: ContactMessage[];
  total: number;
  page: number;
  totalPages: number;
}