# lib/api-zod/src/generated/types/updateCareerApplicationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enitan Afolabi & Company API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCareerApplicationInputStatus } from "./updateCareerApplicationInputStatus";

export interface UpdateCareerApplicationInput {
  status?: UpdateCareerApplicationInputStatus;
}