# lib/api-zod/src/generated/types/updateContactMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enitan Afolabi & Company API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateContactMessageInputStatus } from "./updateContactMessageInputStatus";

export interface UpdateContactMessageInput {
  status?: UpdateContactMessageInputStatus;
}